requirements.txt
pygame==2.5.2
python-socketio==5.10.0
eventlet==0.33.3
requests==2.31.0